Initial Commit
CoreBot is a minecraft bot coded in javascript. CoreBot uses the
mineflayer, mineflayer-navigate, readline and fs APIs and their
dependencies. CoreBot is a bot that can:
- Talk in chat
- Use commands
- Follow players
- Attack players
- Display its health and food values
- Manage its inventory

CoreBot is meant to be used as a companion, it's not going to do
everything. But it might, someday.

To use the bot, click on the "CoreBot ⭐ Alpha Build 5 (Hail).ink" file
after configurating the "config.json" to your liking. Feel free to fork
and edit to your liking. Please don't try to profit off of the creation,
it's not fully yours.